Task 13 spring data (#18)
* add jpa database

* add fixes for jpa database to work properly

* fix tests to work with jpa, change id type in model classes to long

* fix tests to work with jpa

* delete unnecessary lines in build gradle

+++ src/main/java/pl/futurecollars/invoicing/controller/invoice/InvoiceController.java
@@ -22,7 +22,7 @@ public class InvoiceController implements InvoiceApi {
     }
 
     @Override
-    public int save(@RequestBody Invoice invoice) {
+    public long save(@RequestBody Invoice invoice) {
         return invoiceService.save(invoice);
     }
