+++ src/main/java/pl/futurecollars/invoicing/db/memory/InMemoryDatabase.java
@@ -10,18 +10,19 @@ import pl.futurecollars.invoicing.model.Invoice;
 public class InMemoryDatabase implements Database {
 
     private final HashMap<Integer, Invoice> invoices = new HashMap<>();
-    private int index = 1;
+    private long index = 1;
 
     @Override
-    public int save(Invoice invoice) {
+    public long save(Invoice invoice) {
         invoice.setId(index);
-        invoices.put(index, invoice);
+        invoices.put((int) index, invoice);
         return index++;
     }
 
     @Override
-    public Optional<Invoice> getById(int id) {
-        return Optional.ofNullable(invoices.get(id));
+    public Optional<Invoice> getById(long id) {
+
+        return Optional.ofNullable(invoices.get((int) id));
     }
 
     @Override
@@ -31,15 +32,15 @@ public class InMemoryDatabase implements Database {
     }
 
     @Override
-    public Optional<Invoice> update(int id, Invoice updatedInvoice) {
+    public Optional<Invoice> update(long id, Invoice updatedInvoice) {
         updatedInvoice.setId(id);
-        return Optional.ofNullable(invoices.put(id, updatedInvoice));
+        return Optional.ofNullable(invoices.put((int) id, updatedInvoice));
 
     }
 
     @Override
-    public Optional<Invoice> delete(int id) {
-        return Optional.ofNullable(invoices.remove(id));
+    public Optional<Invoice> delete(long id) {
+        return Optional.ofNullable(invoices.remove((int) id));
 
     }
